+++ src/Promotion/Contracts/PromotionEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Promotion\Contracts;
+
+interface PromotionEvent
+{
+    public function getPromotion(): Promotion;
+}